requirements.txt
requests==2.31.0
aiohttp==3.8.5
pytest-asyncio==0.21.1
